refactor(config): move permissions toml types to dedicated module
